sentry upgrade due to xcode update
