+++ package.json
@@ -83,7 +83,7 @@
     "@reown/appkit-auth-wagmi-react-native": "1.0.2",
     "@reown/appkit-wagmi-react-native": "1.0.2",
     "@reown/walletkit": "^1.0.0",
-    "@sentry/react-native": "^5.33.1",
+    "@sentry/react-native": "^7.1.0",
     "@solana/spl-token": "^0.4.8",
     "@solana/web3.js": "^1.98.0",
     "@tanstack/react-query": "^5.56.2",
@@ -154,6 +154,7 @@
     "react-native-camera": "^4.2.1",
     "react-native-collapsible": "^1.6.0",
     "react-native-config": "^1.5.5",
+    "react-native-crypto": "^2.2.1",
     "react-native-custom-qr-codes": "^2.0.0",
     "react-native-device-info": "^8.7.1",
     "react-native-element-dropdown": "^1.8.13",
